requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
attrs==19.3.0
click==7.0
jinja2==2.10.3
markupsafe==1.1.1         # via jinja2
